# backend/openapi/swagger.yaml
basePath: /
definitions:
...
  domain.ObjectUploadResp:
    properties:
      filename:
        type: string
      key:
        type: string
    type: object
...
info:
  contact: {}
  description: niuniu-wiki API documentation
  title: niuniu-wiki API
  version: "1.0"
paths:
...
  /api/v1/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload File
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Knowledge Base ID
        in: formData
        name: kb_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ObjectUploadResp'
      summary: Upload File
      tags:
      - file
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"